as classes singleton foram corrigidas e as classes de teste dos repositorios foram criadas
