+++ Cod3rsGrowth.Infra/Singletons/JogoSingleton.cs
@@ -7,16 +7,17 @@ namespace Cod3rsGrowth.Infra.Singletons
         public JogoSingleton() { }
 
         private static JogoSingleton? _instancia;
-        private static Object _singletonLock = new Object();
 
         public static JogoSingleton Instancia
         {
             get
             {
-                lock (_singletonLock)
+                lock (typeof(JogoSingleton))
                 {
-                    if (_instancia == null) 
-                        _instancia = new JogoSingleton();
+                    if (_instancia == null)
+                    {
+                        _instancia = new JogoSingleton ();
+                    }
                 }
                 return _instancia;
             }
