+++ Cod3rsGrowth.Testes/Mocks/JogoRepositorioMock.cs
@@ -1,6 +1,5 @@
 ﻿using Cod3rsGrowth.Dominio.Entidades;
 using Cod3rsGrowth.Infra.Interfaces;
-using Cod3rsGrowth.Infra.Singletons;
 
 namespace Cod3rsGrowth.Testes.Mocks
 {
@@ -29,36 +28,6 @@ namespace Cod3rsGrowth.Testes.Mocks
         public List<Jogo> ObterTodos()
         {
             throw new NotImplementedException();
-        }
-
-        public void CriarLista()
-        {
-            var listaDeJogo = new List<Jogo>
-            {
-                new Jogo
-                {
-                    Id = 1,
-                    Nome = "Minecraft",
-                    Genero = Dominio.EnumGenero.Genero.SOBREVIVENCIA,
-                    Preco = 100m
-                },
-                new Jogo
-                {
-                    Id = 2,
-                    Nome = "GTA",
-                    Genero = Dominio.EnumGenero.Genero.TPS,
-                    Preco = 200m
-                },
-                new Jogo
-                {
-                    Id = 3,
-                    Nome = "Counter Strike 2",
-                    Genero = Dominio.EnumGenero.Genero.FPS,
-                    Preco = 60m
-                }
-            };
-
-            JogoSingleton.Instancia.AddRange(listaDeJogo);
         }
     }
 }
