+++ Cod3rsGrowth.Testes/Mocks/TesteDeJogoRepositorioMock.cs
@@ -1,6 +1,5 @@
 ﻿using Cod3rsGrowth.Dominio.Entidades;
 using Cod3rsGrowth.Infra.Interfaces;
-using Cod3rsGrowth.Infra.Singletons;
 
 namespace Cod3rsGrowth.Testes.Mocks
 {
@@ -29,45 +28,6 @@ namespace Cod3rsGrowth.Testes.Mocks
         public List<TesteDeJogo> ObterTodos()
         {
             throw new NotImplementedException();
-        }
-
-        public void CriarLista()
-        {
-            var listaDeTesteDeJogo = new List<TesteDeJogo>
-            {
-                new TesteDeJogo
-                {
-                    Id = 1,
-                    NomeResponsavelDoTeste = "Rafael",
-                    Descricao = "O jogo é top",
-                    Nota = 9m,
-                    Aprovado = true,
-                    DataRealizacaoTeste = DateTime.Parse("22/05/2024"),
-                    JogoId = 1
-                },
-                new TesteDeJogo
-                {
-                    Id = 2,
-                    NomeResponsavelDoTeste = "Paulo",
-                    Descricao = "Não gostei do jogo",
-                    Nota = 4.5m,
-                    Aprovado = false,
-                    DataRealizacaoTeste = DateTime.Parse("10/04/2024"),
-                    JogoId = 2
-                },
-                new TesteDeJogo
-                {
-                    Id = 3,
-                    NomeResponsavelDoTeste = "Italo",
-                    Descricao = "Não é um jogo perfeito, mas é jogável",
-                    Nota = 7.4m,
-                    Aprovado = true,
-                    DataRealizacaoTeste = DateTime.Parse("15/06/2024"),
-                    JogoId = 3
-                }
-            };
-
-            TesteDeJogoSingleton.Instancia.AddRange(listaDeTesteDeJogo);
         }
     }
 }
